@@ -19,7 +19,7 @@ func TestToQuery(t *testing.T) {
 
 	assert.Nil(t, err)
 	assert.Equal(t, "SELECT * FROM user WHERE id = ?", sql)
-	assert.Equal(t, []interface{}{1}, args)
+	assert.Equal(t, []any{1}, args)
 
 	sql, args, err = builder.Wrap(
 		Table("user"),
@@ -28,7 +28,7 @@ func TestToQuery(t *testing.T) {
 
 	assert.Nil(t, err)
 	assert.Equal(t, "SELECT * FROM user WHERE name = ? AND age > ?", sql)
-	assert.Equal(t, []interface{}{"yiigo", 20}, args)
+	assert.Equal(t, []any{"yiigo", 20}, args)
 
 	sql, args, err = builder.Wrap(
 		Table("user"),
@@ -37,7 +37,7 @@ func TestToQuery(t *testing.T) {
 
 	assert.Nil(t, err)
 	assert.Equal(t, "SELECT * FROM user WHERE age IN (?, ?)", sql)
-	assert.Equal(t, []interface{}{20, 30}, args)
+	assert.Equal(t, []any{20, 30}, args)
 
 	sql, args, err = builder.Wrap(
 		Table("user"),
@@ -47,7 +47,7 @@ func TestToQuery(t *testing.T) {
 
 	assert.Nil(t, err)
 	assert.Equal(t, "SELECT id, name, age FROM user WHERE id = ?", sql)
-	assert.Equal(t, []interface{}{1}, args)
+	assert.Equal(t, []any{1}, args)
 
 	sql, args, err = builder.Wrap(
 		Table("user"),
@@ -57,7 +57,7 @@ func TestToQuery(t *testing.T) {
 
 	assert.Nil(t, err)
 	assert.Equal(t, "SELECT DISTINCT name FROM user WHERE id = ?", sql)
-	assert.Equal(t, []interface{}{1}, args)
+	assert.Equal(t, []any{1}, args)
 
 	sql, args, err = builder.Wrap(
 		Table("user"),
@@ -67,7 +67,7 @@ func TestToQuery(t *testing.T) {
 
 	assert.Nil(t, err)
 	assert.Equal(t, "SELECT * FROM user INNER JOIN address ON user.id = address.user_id WHERE user.id = ?", sql)
-	assert.Equal(t, []interface{}{1}, args)
+	assert.Equal(t, []any{1}, args)
 
 	sql, args, err = builder.Wrap(
 		Table("user"),
@@ -77,7 +77,7 @@ func TestToQuery(t *testing.T) {
 
 	assert.Nil(t, err)
 	assert.Equal(t, "SELECT * FROM user LEFT JOIN address ON user.id = address.user_id WHERE user.id = ?", sql)
-	assert.Equal(t, []interface{}{1}, args)
+	assert.Equal(t, []any{1}, args)
 
 	sql, args, err = builder.Wrap(
 		Table("user"),
@@ -87,7 +87,7 @@ func TestToQuery(t *testing.T) {
 
 	assert.Nil(t, err)
 	assert.Equal(t, "SELECT * FROM user RIGHT JOIN address ON user.id = address.user_id WHERE user.id = ?", sql)
-	assert.Equal(t, []interface{}{1}, args)
+	assert.Equal(t, []any{1}, args)
 
 	sql, args, err = builder.Wrap(
 		Table("user"),
@@ -97,7 +97,7 @@ func TestToQuery(t *testing.T) {
 
 	assert.Nil(t, err)
 	assert.Equal(t, "SELECT * FROM user FULL JOIN address ON user.id = address.user_id WHERE user.id = ?", sql)
-	assert.Equal(t, []interface{}{1}, args)
+	assert.Equal(t, []any{1}, args)
 
 	sql, _, err = builder.Wrap(
 		Table("sizes"),
@@ -116,7 +116,7 @@ func TestToQuery(t *testing.T) {
 
 	assert.Nil(t, err)
 	assert.Equal(t, "SELECT * FROM user LEFT JOIN address ON user.id = address.user_id RIGHT JOIN company ON user.id = company.user_id WHERE user.id = ?", sql)
-	assert.Equal(t, []interface{}{1}, args)
+	assert.Equal(t, []any{1}, args)
 
 	sql, args, err = builder.Wrap(
 		Table("address"),
@@ -127,7 +127,7 @@ func TestToQuery(t *testing.T) {
 
 	assert.Nil(t, err)
 	assert.Equal(t, "SELECT user_id, COUNT(*) AS total FROM address GROUP BY user_id HAVING user_id = ?", sql)
-	assert.Equal(t, []interface{}{1}, args)
+	assert.Equal(t, []any{1}, args)
 
 	sql, args, err = builder.Wrap(
 		Table("user"),
@@ -139,7 +139,7 @@ func TestToQuery(t *testing.T) {
 
 	assert.Nil(t, err)
 	assert.Equal(t, "SELECT * FROM user WHERE age > ? ORDER BY age ASC, id DESC LIMIT ? OFFSET ?", sql)
-	assert.Equal(t, []interface{}{20, 10, 5}, args)
+	assert.Equal(t, []any{20, 10, 5}, args)
 
 	sql, args, err = builder.Wrap(
 		Table("user_0"),
@@ -149,7 +149,7 @@ func TestToQuery(t *testing.T) {
 
 	assert.Nil(t, err)
 	assert.Equal(t, "(SELECT * FROM user_0 WHERE id = ?) UNION (SELECT * FROM user_1 WHERE id = ?)", sql)
-	assert.Equal(t, []interface{}{1, 2}, args)
+	assert.Equal(t, []any{1, 2}, args)
 
 	sql, args, err = builder.Wrap(
 		Table("user_0"),
@@ -159,7 +159,7 @@ func TestToQuery(t *testing.T) {
 
 	assert.Nil(t, err)
 	assert.Equal(t, "(SELECT * FROM user_0 WHERE id = ?) UNION ALL (SELECT * FROM user_1 WHERE id = ?)", sql)
-	assert.Equal(t, []interface{}{1, 2}, args)
+	assert.Equal(t, []any{1, 2}, args)
 
 	sql, args, err = builder.Wrap(
 		Table("user_0"),
@@ -176,7 +176,7 @@ func TestToQuery(t *testing.T) {
 
 	assert.Nil(t, err)
 	assert.Equal(t, "(SELECT * FROM user_0 WHERE age IN (?, ?) LIMIT ?) UNION (SELECT * FROM user_1 WHERE age IN (?, ?) LIMIT ?)", sql)
-	assert.Equal(t, []interface{}{10, 20, 5, 30, 40, 5}, args)
+	assert.Equal(t, []any{10, 20, 5, 30, 40, 5}, args)
 
 	sql, args, err = builder.Wrap(
 		Table("user_0"),
@@ -187,7 +187,7 @@ func TestToQuery(t *testing.T) {
 
 	assert.Nil(t, err)
 	assert.Equal(t, "(SELECT * FROM user_0 WHERE id = ?) UNION (SELECT * FROM user_1 WHERE id = ?) UNION ALL (SELECT * FROM user_2 WHERE id = ?)", sql)
-	assert.Equal(t, []interface{}{1, 2, 3}, args)
+	assert.Equal(t, []any{1, 2, 3}, args)
 }
 
 func TestToInsert(t *testing.T) {
@@ -211,7 +211,7 @@ func TestToInsert(t *testing.T) {
 
 	assert.Nil(t, err)
 	assert.Equal(t, "INSERT INTO user (name, gender, age) VALUES (?, ?, ?)", sql)
-	assert.Equal(t, []interface{}{"yiigo", "M", 29}, args)
+	assert.Equal(t, []any{"yiigo", "M", 29}, args)
 
 	sql, args, err = builder.Wrap(Table("user")).ToInsert(ctx, &User{
 		Name:   "yiigo",
@@ -222,7 +222,7 @@ func TestToInsert(t *testing.T) {
 
 	assert.Nil(t, err)
 	assert.Equal(t, "INSERT INTO user (name, gender, age, phone) VALUES (?, ?, ?, ?)", sql)
-	assert.Equal(t, []interface{}{"yiigo", "M", 29, "13605109425"}, args)
+	assert.Equal(t, []any{"yiigo", "M", 29, "13605109425"}, args)
 
 	// map 字段顺序不一定
 	// sql, args, err = builder.Wrap(Table("user")).ToInsert(ctx, X{
@@ -232,7 +232,7 @@ func TestToInsert(t *testing.T) {
 	// })
 	//
 	// assert.Equal(t, "INSERT INTO user (age, gender, name) VALUES (?, ?, ?)", sql)
-	// assert.Equal(t, []interface{}{29, "M", "yiigo"}, args)
+	// assert.Equal(t, []any{29, "M", "yiigo"}, args)
 }
 
 func TestToBatchInsert(t *testing.T) {
@@ -263,7 +263,7 @@ func TestToBatchInsert(t *testing.T) {
 
 	assert.Nil(t, err)
 	assert.Equal(t, "INSERT INTO user (name, gender, age) VALUES (?, ?, ?), (?, ?, ?)", sql)
-	assert.Equal(t, []interface{}{"yiigo", "M", 29, "test", "W", 20}, args)
+	assert.Equal(t, []any{"yiigo", "M", 29, "test", "W", 20}, args)
 
 	sql, args, err = builder.Wrap(Table("user")).ToBatchInsert(ctx, []*User{
 		{
@@ -282,7 +282,7 @@ func TestToBatchInsert(t *testing.T) {
 
 	assert.Nil(t, err)
 	assert.Equal(t, "INSERT INTO user (name, gender, age, phone) VALUES (?, ?, ?, ?), (?, ?, ?, ?)", sql)
-	assert.Equal(t, []interface{}{"yiigo", "M", 29, "13605109425", "test", "W", 20, "13605105471"}, args)
+	assert.Equal(t, []any{"yiigo", "M", 29, "13605109425", "test", "W", 20, "13605105471"}, args)
 
 	// map 字段顺序不一定
 	// sql, args, err = builder.Wrap(Table("user")).ToBatchInsert(ctx, []X{
@@ -299,7 +299,7 @@ func TestToBatchInsert(t *testing.T) {
 	// })
 	//
 	// assert.Equal(t, "INSERT INTO user (age, gender, name) VALUES (?, ?, ?), (?, ?, ?)", sql)
-	// assert.Equal(t, []interface{}{29, "M", "yiigo", 20, "W", "test"}, args)
+	// assert.Equal(t, []any{29, "M", "yiigo", 20, "W", "test"}, args)
 }
 
 func TestToUpdate(t *testing.T) {
@@ -325,7 +325,7 @@ func TestToUpdate(t *testing.T) {
 
 	assert.Nil(t, err)
 	assert.Equal(t, "UPDATE user SET name = ?, gender = ?, age = ? WHERE id = ?", sql)
-	assert.Equal(t, []interface{}{"yiigo", "M", 29, 1}, args)
+	assert.Equal(t, []any{"yiigo", "M", 29, 1}, args)
 
 	sql, args, err = builder.Wrap(
 		Table("user"),
@@ -339,7 +339,7 @@ func TestToUpdate(t *testing.T) {
 
 	assert.Nil(t, err)
 	assert.Equal(t, "UPDATE user SET name = ?, gender = ?, age = ?, phone = ? WHERE id = ?", sql)
-	assert.Equal(t, []interface{}{"yiigo", "M", 29, "13605109425", 1}, args)
+	assert.Equal(t, []any{"yiigo", "M", 29, "13605109425", 1}, args)
 
 	// map 字段顺序不一定
 	// sql, args, err = builder.Wrap(
@@ -352,7 +352,7 @@ func TestToUpdate(t *testing.T) {
 	// })
 	//
 	// assert.Equal(t, "UPDATE user SET age = ?, gender = ?, name = ? WHERE id = ?", sql)
-	// assert.Equal(t, []interface{}{29, "M", "yiigo", 1}, args)
+	// assert.Equal(t, []any{29, "M", "yiigo", 1}, args)
 
 	sql, args, err = builder.Wrap(
 		Table("user"),
@@ -365,7 +365,7 @@ func TestToUpdate(t *testing.T) {
 
 	assert.Nil(t, err)
 	assert.Equal(t, "UPDATE user SET name = ?, gender = ?, age = ? WHERE id IN (?, ?)", sql)
-	assert.Equal(t, []interface{}{"yiigo", "M", 29, 1, 2}, args)
+	assert.Equal(t, []any{"yiigo", "M", 29, 1, 2}, args)
 
 	sql, args, err = builder.Wrap(
 		Table("product"),
@@ -374,7 +374,7 @@ func TestToUpdate(t *testing.T) {
 
 	assert.Nil(t, err)
 	assert.Equal(t, "UPDATE product SET price = price * ? + ? WHERE id = ?", sql)
-	assert.Equal(t, []interface{}{2, 100, 1}, args)
+	assert.Equal(t, []any{2, 100, 1}, args)
 }
 
 func TestToDelete(t *testing.T) {
@@ -389,7 +389,7 @@ func TestToDelete(t *testing.T) {
 
 	assert.Nil(t, err)
 	assert.Equal(t, "DELETE FROM user WHERE id = ?", sql)
-	assert.Equal(t, []interface{}{1}, args)
+	assert.Equal(t, []any{1}, args)
 
 	sql, args, err = builder.Wrap(
 		Table("user"),
@@ -398,7 +398,7 @@ func TestToDelete(t *testing.T) {
 
 	assert.Nil(t, err)
 	assert.Equal(t, "DELETE FROM user WHERE id IN (?, ?)", sql)
-	assert.Equal(t, []interface{}{1, 2}, args)
+	assert.Equal(t, []any{1, 2}, args)
 }
 
 func TestToTruncate(t *testing.T) {
